+++ src/main/kotlin/hiiragi283/ragium/common/storage/item/HTPotionBundleItemHandler.kt
@@ -7,7 +7,7 @@ import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.Items
 
-class HTPotionBundleItemHandler(parent: ItemStack, size: Int) :
+class HTPotionBundleItemHandler(parent: ItemStack, size: Long) :
     HTComponentItemHandler(parent, size),
     HTMenuCallback {
     companion object {
